change func name

+++ features/user/home_account/src/main/java/com/tokopedia/home_account/view/mapper/DataViewMapper.kt
@@ -103,7 +103,7 @@ class DataViewMapper @Inject constructor(
         val totalBalance = balance.buyerUsable + balance.sellerUsable
         return CommonDataView(
                 id = AccountConstants.SettingCode.SETTING_SALDO,
-                title = CurrencyFormatUtil.convertPriceValueToIdrFormat(totalBalance.toLong() , false),
+                title = CurrencyFormatUtil.convertPriceValueToIdrFormat(totalBalance.toDouble(), false),
                 body = getString(context, R.string.account_title_saldo_item),
                 type = FinancialItemViewHolder.TYPE_SALDO,
                 urlIcon = AccountConstants.Url.SALDO_ICON,
